Moving stuff around into submodules; Refactoring of imports
