+++ clay/core/__init__.py
@@ -1,0 +1,6 @@
+from .eventhook import EventHook
+from .gp import gp
+from .log import logger
+from .player import player
+from .settings import settings_manager
+from .osd import osd_manager
