+++ clay/ui/urwid/pages/__init__.py
@@ -1,0 +1,8 @@
+
+from .debug import DebugPage
+from .library import LibraryPage
+from .playlists import PlaylistsPage
+from .stations import StationsPage
+from .playerqueue import QueuePage
+from .search import SearchPage
+from .settings import SettingsPage
